feat: 즉시 거래 로직 및 주문 생성 기능 - #39
- 즉시 구매/판매 실행을 위한 MatchInstantTradeUseCase 구현
- OrderMapper 추가 및 주문(Order) 엔티티 매핑 로직 작성
- BIDDING_NOT_FOUND 에러코드 정의 및 예외 처리 로직 반영
- 통합 테스트 케이스 작성하여 주요 시나리오 검증
- OrderRepository, Bidding 도메인에 상태 변경 메서드 추가

+++ market/src/main/java/com/back/market/domain/Order.java
@@ -11,6 +11,7 @@ import lombok.experimental.SuperBuilder;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.SQLRestriction;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -36,7 +37,7 @@ public class Order extends BaseTimeEntity {
     private Bidding sellBidding;                // 판매 입찰 내역(1:1)
 
     @Column(name = "price", nullable = false)
-    private Long price;                          // 체결 가격(입찰가 변동과 무관)
+    private BigDecimal price;                          // 체결 가격(입찰가 변동과 무관)
 
     @Column(name = "address", nullable = false, length = 500)
     private String address;                      // 배송지 주소(사용자 주소 변경과 무관)
